Package template
Creating a package output template. Porting the
https://github.com/ngi-nix/gnunet-messenger-cli flake as an example.

Re-using the Nixpkgs call-package scope to build the packageset. We
then expose this newly packagset on the flake boundary.

We don't leverage the flake inputs for the packages sources. I'm a bit
worried it'd lead to a lot of lockfile-fueled conflicts down the road.
I'm not 100% convinced this is the right approach though, we might
want to reconsider this down the road.

This involves a bit of boilerplate to add the package entry to the
all-packages file. We could probably retrieve the packages from the
repository tree. However, I feel like this explicit approach involves
less magic and is simpler to understand.