Reuse ssh-keys for non-critical-infra
Creates the `infra` key list and migrates hetzner-01 to run with it.

The wrong keys were deployed for me, so I had to access the host using
the keypair I use for GitHub.

Fetching keys from GitHub is not great for key hygiene and did in fact
cause a hash mismatch on Julien's keys.