atomic-server: refactor to rebuild from github source
This refactor builds atomic-server using the separately packaged
atomic-browser and then the github source. Not only is this more
reflective of the real build process used, but it is helpful when trying
to debug atomic-server by building a fork with additional logging, which
we did while working on https://github.com/ngi-nix/ngipkgs/pull/220

Previously we were building atomic-server from the published crate,
which is usually "from source". But in this case that meant only the
rust source, and separately built web assets were packaged into the
crate. This can be verified by running
`curl -L 'https://crates.io/api/v1/crates/atomic-server/0.37.0/download' | tar -tv`
and checking for the `assets_tmp` directory.