flake.nix: Drop systemAgnosticOutputs.nixosConfigurations
There was a TODO about removing them, and it caused issues now when we evaluated into them
and got to parsing `system = null`.

`makemake` has been moved into `eachDefaultSystemOutputs.nixosConfigurations`.
Nixpkgs was then not happy with the way we were importing the system configuration for `makemake`,
so adjust the entry point of our import until it's happy.

- Errored out on us importing an entire config into the module system, so change it into a module
- sops-nix was getting imported twice (it's in `default.nix`'s `extendedNixosModules` already),
  so drop the second import from `makemake`.