heads: Assert that specified board actually exists
Invalid/No-longer-existing board names have so far just been silently getting filtered out of the resulting attrset.
This allowed some typos in board names to fly under the radar, so clearly this wasn't a great idea.

Drop the silent filtering, assert that the boards exist. Also assert on the list of known-broken boards, and the
board settings overrides attrset names.